feat: Add option to compress IP ranges
When blocking too many IPs, an error
> Hash is full, cannot add more elements
might be shown and the service fails to activate. This can easily happen
when blocking many ASes like [AS7552](https://bgp.tools/as/7552) which
announces almost 8k routes. That quickly fills our hashset. On closer
inspection, there are many routes, but they are all within a few much
bigger IP ranges. Since we do not care about the individual routes, but
only the completed prefixes, we can just merge them.

Since it is much easier to achieve this in python, it would add a
dependency and also this consumes much more memory while processing the
IPs and so this is not enabled by default.